+++ database/point.go
@@ -127,8 +127,9 @@ func (d *GormDatabase) CreatePoint(body *model.Point, fromPlugin bool) (*model.P
 	return body, nil
 }
 
-func (d *GormDatabase) UpdatePoint(uuid string, body *model.Point, fromPlugin bool, afterRealDeviceUpdate bool) (
+func (d *GormDatabase) UpdatePoint(uuid string, body *model.Point, fromPlugin bool) (
 	*model.Point, error) {
+	writeOnDB := !fromPlugin
 	var pointModel *model.Point
 	query := d.DB.Where("uuid = ?", uuid).Preload("Tags").Preload("MetaTags").
 		Preload("Priority").First(&pointModel)
@@ -150,13 +151,13 @@ func (d *GormDatabase) UpdatePoint(uuid string, body *model.Point, fromPlugin bo
 			return nil, errors.New(eMsg)
 		}
 	}
-	if len(body.Tags) > 0 && !fromPlugin {
+	if len(body.Tags) > 0 && writeOnDB {
 		if err := d.updateTags(&pointModel, body.Tags); err != nil {
 			return nil, err
 		}
 	}
 	publishPointList := body.Name != pointModel.Name
-	if !fromPlugin {
+	if writeOnDB {
 		if err := d.DB.Model(&pointModel).Select("*").Updates(&body).Error; err != nil {
 			return nil, err
 		}
@@ -169,8 +170,8 @@ func (d *GormDatabase) UpdatePoint(uuid string, body *model.Point, fromPlugin bo
 		pointModel.Priority = body.Priority
 	}
 	priorityMap := priorityarray.ConvertToMap(*pointModel.Priority)
-	pnt, _, _, _, err := d.updatePointValue(pointModel, &priorityMap, fromPlugin, afterRealDeviceUpdate, nil, false)
-	if !fromPlugin {
+	pnt, _, _, _, err := d.updatePointValue(pointModel, &priorityMap, nil, false, writeOnDB)
+	if writeOnDB {
 		if publishPointList {
 			go d.PublishPointsList("")
 		}
@@ -181,14 +182,13 @@ func (d *GormDatabase) UpdatePoint(uuid string, body *model.Point, fromPlugin bo
 			return nil, err
 		}
 	} else {
-		d.bufferPointUpdate(uuid, body, afterRealDeviceUpdate)
+		d.bufferPointUpdate(uuid, body)
 	}
 	return pnt, err
 }
 
-func (d *GormDatabase) PointWrite(uuid string, body *model.PointWriter, fromPlugin bool, afterRealDeviceUpdate bool,
-	currentWriterUUID *string, forceWrite bool) (returnPoint *model.Point, isPresentValueChange, isWriteValueChange,
-	isPriorityChanged bool, err error) {
+func (d *GormDatabase) PointWrite(uuid string, body *model.PointWriter, currentWriterUUID *string, forceWrite bool) (
+	returnPoint *model.Point, isPresentValueChange, isWriteValueChange, isPriorityChanged bool, err error) {
 	var pointModel *model.Point
 	query := d.DB.Where("uuid = ?", uuid).Preload("Priority").First(&pointModel)
 	if query.Error != nil {
@@ -201,21 +201,17 @@ func (d *GormDatabase) PointWrite(uuid string, body *model.PointWriter, fromPlug
 		pointModel.ValueUpdatedFlag = boolean.NewTrue()
 	}
 	point, isPresentValueChange, isWriteValueChange, isPriorityChanged, err :=
-		d.updatePointValue(pointModel, body.Priority, fromPlugin, afterRealDeviceUpdate, currentWriterUUID, forceWrite)
-	if fromPlugin {
-		d.bufferPointWrite(uuid, body, afterRealDeviceUpdate, currentWriterUUID, forceWrite)
-	}
+		d.updatePointValue(pointModel, body.Priority, currentWriterUUID, forceWrite, true)
 	return point, isPresentValueChange, isWriteValueChange, isPriorityChanged, err
 }
 
-func (d *GormDatabase) updatePointValue(pointModel *model.Point, priority *map[string]*float64, fromPlugin bool,
-	afterRealDeviceUpdate bool, currentWriterUUID *string, forceWrite bool) (returnPoint *model.Point,
-	isPresentValueChange, isWriteValueChange, isPriorityChanged bool, err error) {
+func (d *GormDatabase) updatePointValue(pointModel *model.Point, priority *map[string]*float64, currentWriterUUID *string, forceWrite, writeOnDB bool) (
+	returnPoint *model.Point, isPresentValueChange, isWriteValueChange, isPriorityChanged bool, err error) {
 	if pointModel.PointPriorityArrayMode == "" {
 		pointModel.PointPriorityArrayMode = model.PriorityArrayToPresentValue // sets default priority array mode
 	}
 
-	pointModel, priority, presentValue, writeValue, isPriorityChanged := d.updatePriority(pointModel, priority, fromPlugin)
+	pointModel, priority, presentValue, writeValue, isPriorityChanged := d.updatePriority(pointModel, priority, writeOnDB)
 	ov := float.Copy(presentValue)
 	pointModel.OriginalValue = ov
 	wv := float.Copy(writeValue)
@@ -224,13 +220,6 @@ func (d *GormDatabase) updatePointValue(pointModel *model.Point, priority *map[s
 	presentValueTransformFault := false
 	transform := PointValueTransformOnRead(presentValue, pointModel.ScaleEnable, pointModel.MultiplicationFactor,
 		pointModel.ScaleInMin, pointModel.ScaleInMax, pointModel.ScaleOutMin, pointModel.ScaleOutMax, pointModel.Offset)
-	if afterRealDeviceUpdate {
-		pointModel.CommonFault.InFault = false
-		pointModel.CommonFault.MessageLevel = model.MessageLevel.Info
-		pointModel.CommonFault.MessageCode = model.CommonFaultCode.PointWriteOk
-		pointModel.CommonFault.Message = fmt.Sprintf("last-updated: %s", utilstime.TimeStamp())
-		pointModel.CommonFault.LastOk = time.Now().UTC()
-	}
 	presentValue = transform
 	val, err := pointUnits(presentValue, pointModel.Unit, pointModel.UnitTo)
 	if err != nil {
@@ -244,17 +233,6 @@ func (d *GormDatabase) updatePointValue(pointModel *model.Point, priority *map[s
 		presentValue = val
 	}
 	writeValue = PointValueTransformOnWrite(writeValue, pointModel.ScaleEnable, pointModel.MultiplicationFactor, pointModel.ScaleInMin, pointModel.ScaleInMax, pointModel.ScaleOutMin, pointModel.ScaleOutMax, pointModel.Offset)
-
-	// example for wires and modbus:
-	// if a new value is written from wires then set this to false so the modbus knows on the next poll to write a new
-	// value to the modbus point
-	if afterRealDeviceUpdate {
-		pointModel.InSync = boolean.NewTrue() // TODO: do we still use InSync?
-		// pointModel.WritePollRequired = boolean.NewFalse()  // WritePollRequired should be set by the plugins (they know best)
-	} else {
-		pointModel.InSync = boolean.NewFalse() // TODO: do we still use InSync?
-		// pointModel.WritePollRequired = boolean.NewTrue()  // WritePollRequired should be set by the plugins (they know best)
-	}
 
 	if !integer.IsUnit32Nil(pointModel.Decimal) && presentValue != nil {
 		value := nmath.RoundTo(*presentValue, *pointModel.Decimal)
@@ -281,10 +259,10 @@ func (d *GormDatabase) updatePointValue(pointModel *model.Point, priority *map[s
 	pointModel.MessageCode = model.CommonFaultCode.Ok
 	pointModel.Message = fmt.Sprintf("lastMessage: %s", utilstime.TimeStamp())
 	pointModel.LastOk = time.Now()
-	if !fromPlugin {
+	if !writeOnDB {
 		_ = d.DB.Model(&pointModel).Select("*").Updates(&pointModel)
 	}
-	if isChange && !fromPlugin {
+	if isChange && !writeOnDB {
 		err = d.ProducersPointWrite(pointModel.UUID, priority, pointModel.PresentValue, isPresentValueChange,
 			currentWriterUUID)
 		if err != nil {
@@ -365,13 +343,12 @@ func (d *GormDatabase) DeletePoint(uuid string) (bool, error) {
 	return r != 0, nil
 }
 
-func (d *GormDatabase) PointWriteByName(networkName, deviceName, pointName string, body *model.PointWriter,
-	fromPlugin bool) (*model.Point, error) {
+func (d *GormDatabase) PointWriteByName(networkName, deviceName, pointName string, body *model.PointWriter) (*model.Point, error) {
 	point, err := d.GetPointByName(networkName, deviceName, pointName, api.Args{})
 	if err != nil {
 		return nil, err
 	}
-	write, _, _, _, err := d.PointWrite(point.UUID, body, fromPlugin, false, nil, false)
+	write, _, _, _, err := d.PointWrite(point.UUID, body, nil, false)
 	if err != nil {
 		return nil, err
 	}