+++ database/pointmqtt.go
@@ -47,7 +47,7 @@ func (d *GormDatabase) publishPointWrite(details *interfaces.MqttPoint) {
 		return
 	}
 	if details.PointUUID != "" {
-		_, _, _, _, err := d.PointWrite(details.PointUUID, details.Priority, false, false, nil, false)
+		_, _, _, _, err := d.PointWrite(details.PointUUID, details.Priority, nil, false)
 		if err != nil {
 			log.Error("mqtt write point by uuid: error:", err)
 			return
@@ -56,7 +56,7 @@ func (d *GormDatabase) publishPointWrite(details *interfaces.MqttPoint) {
 		networkName := details.NetworkName
 		deviceName := details.DeviceName
 		pointName := details.PointName
-		_, err := d.PointWriteByName(networkName, deviceName, pointName, details.Priority, false)
+		_, err := d.PointWriteByName(networkName, deviceName, pointName, details.Priority)
 		if err != nil {
 			log.Error("mqtt write point by name: error:", err)
 			return