+++ interfaces/pointupdatebuffer.go
@@ -3,7 +3,6 @@ package interfaces
 import "github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 
 type PointUpdateBuffer struct {
-	UUID                  string       `json:"uuid"`
-	Body                  *model.Point `json:"body"`
-	AfterRealDeviceUpdate bool         `json:"after_real_device_update"`
+	UUID string       `json:"uuid"`
+	Body *model.Point `json:"body"`
 }