+++ plugin/nube/protocals/lora/api.go
@@ -67,13 +67,13 @@ func (inst *Instance) RegisterWebhook(basePath string, mux *gin.RouterGroup) {
 	})
 	mux.PATCH(plugin.PointsURL, func(ctx *gin.Context) {
 		body, _ := plugin.GetBODYPoint(ctx)
-		point, err := inst.db.UpdatePoint(body.UUID, body, false)
+		point, err := inst.db.UpdatePoint(body.UUID, body)
 		api.ResponseHandler(point, err, ctx)
 	})
 	mux.PATCH(plugin.PointsWriteURL, func(ctx *gin.Context) {
 		body, _ := plugin.GetBodyPointWriter(ctx)
 		uuid := plugin.ResolveID(ctx)
-		point, _, _, _, err := inst.db.PointWrite(uuid, body, false)
+		point, _, _, _, err := inst.db.PointWrite(uuid, body)
 		api.ResponseHandler(point, err, ctx)
 	})
 	mux.DELETE(plugin.NetworksURL, func(ctx *gin.Context) {